- Renaming - Local Config (no webhook yet) - running with local rabbit
