--- accounts/src/main/java/com/eazybytes/accounts/dto/AccountsContactInfoDto.java
@@ -1,19 +1,0 @@
-package com.eazybytes.accounts.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-import java.util.List;
-import java.util.Map;
-
-@ConfigurationProperties(prefix = "accounts")
-@Getter
-@Setter
-public class AccountsContactInfoDto {
-
-    private String message;
-    private Map<String, String> contactDetails;
-    private List<String> onCallSupport;
-
-}
