--- accounts/src/main/java/com/eazybytes/accounts/dto/ErrorResponseDto.java
@@ -1,37 +1,0 @@
-package com.eazybytes.accounts.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import org.springframework.http.HttpStatus;
-
-import java.time.LocalDateTime;
-
-@Data @AllArgsConstructor
-@Schema(
-        name = "ErrorResponse",
-        description = "Schema to hold error response information"
-)
-public class ErrorResponseDto {
-
-    @Schema(
-            description = "API path invoked by client"
-    )
-    private  String apiPath;
-
-    @Schema(
-            description = "Error code representing the error happened"
-    )
-    private HttpStatus errorCode;
-
-    @Schema(
-            description = "Error message representing the error happened"
-    )
-    private  String errorMessage;
-
-    @Schema(
-            description = "Time representing when the error happened"
-    )
-    private LocalDateTime errorTime;
-
-}
