--- accounts/src/main/java/com/eazybytes/accounts/entity/Customer.java
@@ -1,26 +1,0 @@
-package com.eazybytes.accounts.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-public class Customer extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "customer_id")
-    private Long customerId;
-
-    private String name;
-
-    private String email;
-
-    @Column(name = "mobile_number")
-    private String mobileNumber;
-
-}
