--- accounts/src/main/java/com/eazybytes/accounts/exception/CustomerAlreadyExistsException.java
@@ -1,13 +1,0 @@
-package com.eazybytes.accounts.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class CustomerAlreadyExistsException extends RuntimeException {
-
-    public CustomerAlreadyExistsException(String message) {
-        super(message);
-    }
-
-}
