--- accounts/src/main/java/com/eazybytes/accounts/mapper/CustomerMapper.java
@@ -1,22 +1,0 @@
-package com.eazybytes.accounts.mapper;
-
-import com.eazybytes.accounts.dto.CustomerDto;
-import com.eazybytes.accounts.entity.Customer;
-
-public class CustomerMapper {
-
-    public static CustomerDto mapToCustomerDto(Customer customer, CustomerDto customerDto) {
-        customerDto.setName(customer.getName());
-        customerDto.setEmail(customer.getEmail());
-        customerDto.setMobileNumber(customer.getMobileNumber());
-        return customerDto;
-    }
-
-    public static Customer mapToCustomer(CustomerDto customerDto, Customer customer) {
-        customer.setName(customerDto.getName());
-        customer.setEmail(customerDto.getEmail());
-        customer.setMobileNumber(customerDto.getMobileNumber());
-        return customer;
-    }
-
-}
