--- accounts/src/main/java/com/eazybytes/accounts/repository/AccountsRepository.java
@@ -1,20 +1,0 @@
-package com.eazybytes.accounts.repository;
-
-import com.eazybytes.accounts.entity.Accounts;
-import jakarta.transaction.Transactional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface AccountsRepository extends JpaRepository<Accounts, Long> {
-
-    Optional<Accounts> findByCustomerId(Long customerId);
-
-    @Transactional
-    @Modifying
-    void deleteByCustomerId(Long customerId);
-
-}
