--- accounts/src/main/java/com/eazybytes/accounts/repository/CustomerRepository.java
@@ -1,15 +1,0 @@
-package com.eazybytes.accounts.repository;
-
-import com.eazybytes.accounts.entity.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-
-    Optional<Customer> findByMobileNumber(String mobileNumber);
-
-
-}
