+++ accounts/src/main/java/com/nisbank/accounts/controller/AccountsController.java
@@ -1,0 +1,258 @@
+package com.nisbank.accounts.controller;
+
+import com.nisbank.accounts.constants.AccountsConstants;
+import com.nisbank.accounts.dto.AccountContactInfoDto;
+import com.nisbank.accounts.dto.CustomerDto;
+import com.nisbank.accounts.dto.ErrorResponseDto;
+import com.nisbank.accounts.dto.ResponseDto;
+import com.nisbank.accounts.service.IAccountsService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Pattern;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.env.Environment;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+
+import static com.nisbank.accounts.constants.AccountsConstants.*;
+
+/**
+ * @author Nir Bar Levav
+ */
+
+@Tag(
+        name = "Accounts in NisBank",
+        description = "CRUD REST APIs in NisBank to CREATE, UPDATE, FETCH AND DELETE account details"
+)
+@RestController
+@RequestMapping(path="/api", produces = {MediaType.APPLICATION_JSON_VALUE})
+@Validated
+public class AccountsController {
+
+    private final IAccountsService iAccountsService;
+
+    public AccountsController(IAccountsService iAccountsService) {
+        this.iAccountsService = iAccountsService;
+    }
+
+    @Value("${build.version}")
+    private String buildVersion;
+
+    @Autowired
+    private Environment environment;
+
+    @Autowired
+    private AccountContactInfoDto accountContactInfoDto;
+
+    @Operation(
+            summary = "Create Account",
+            description = "REST API to create new Customer & Account inside NisBank"
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = STATUS_201,
+                    description = MESSAGE_201
+            ),
+            @ApiResponse(
+                    responseCode = STATUS_500,
+                    description = MESSAGE_500,
+                    content = @Content(
+                            schema = @Schema(implementation = ErrorResponseDto.class)
+                    )
+            )
+    }
+    )
+    @PostMapping("/create")
+    public ResponseEntity<ResponseDto> createAccount(@Valid @RequestBody CustomerDto customerDto) {
+        iAccountsService.createAccount(customerDto);
+        return ResponseEntity
+                .status(HttpStatus.CREATED)
+                .body(new ResponseDto(STATUS_201, AccountsConstants.MESSAGE_201));
+    }
+
+    @Operation(
+            summary = "Fetch Account Details REST API",
+            description = "REST API to fetch Customer & Account details based on a mobile number"
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = STATUS_200,
+                    description = MESSAGE_200
+            ),
+            @ApiResponse(
+                    responseCode = STATUS_500,
+                    description = MESSAGE_500,
+                    content = @Content(
+                            schema = @Schema(implementation = ErrorResponseDto.class)
+                    )
+            )
+    }
+    )
+    @GetMapping("/fetch")
+    public ResponseEntity<CustomerDto> fetchAccountDetails(@RequestParam
+                                                               @Pattern(regexp="(^$|[0-9]{10})",message = "Mobile number must be 10 digits")
+                                                               String mobileNumber) {
+        CustomerDto customerDto = iAccountsService.fetchAccount(mobileNumber);
+        return ResponseEntity.status(HttpStatus.OK).body(customerDto);
+    }
+
+    @Operation(
+            summary = "Update Account Details REST API",
+            description = "REST API to update Customer & Account details based on a account number"
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = STATUS_200,
+                    description = MESSAGE_200
+            ),
+            @ApiResponse(
+                    responseCode = STATUS_417,
+                    description = MESSAGE_417_UPDATE
+            ),
+            @ApiResponse(
+                    responseCode = STATUS_500,
+                    description = MESSAGE_500,
+                    content = @Content(
+                            schema = @Schema(implementation = ErrorResponseDto.class)
+                    )
+            )
+    }
+    )
+    @PutMapping("/update")
+    public ResponseEntity<ResponseDto> updateAccountDetails(@Valid @RequestBody CustomerDto customerDto) {
+        boolean isUpdated = iAccountsService.updateAccount(customerDto);
+        if(isUpdated) {
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(new ResponseDto(AccountsConstants.STATUS_200, AccountsConstants.MESSAGE_200));
+        }else{
+            return ResponseEntity
+                    .status(HttpStatus.EXPECTATION_FAILED)
+                    .body(new ResponseDto(AccountsConstants.STATUS_417, AccountsConstants.MESSAGE_417_UPDATE));
+        }
+    }
+
+    @Operation(
+            summary = "Delete Account & Customer Details REST API",
+            description = "REST API to delete Customer & Account details based on a mobile number"
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = STATUS_200,
+                    description = MESSAGE_200
+            ),
+            @ApiResponse(
+                    responseCode = STATUS_417,
+                    description = MESSAGE_417_DELETE
+            ),
+            @ApiResponse(
+                    responseCode = STATUS_500,
+                    description = MESSAGE_500,
+                    content = @Content(
+                            schema = @Schema(implementation = ErrorResponseDto.class)
+                    )
+            )
+    }
+    )
+    @DeleteMapping("/delete")
+    public ResponseEntity<ResponseDto> deleteAccountDetails(@RequestParam
+                                                                @Pattern(regexp="(^$|[0-9]{10})",message = "Mobile number must be 10 digits")
+                                                                String mobileNumber) {
+        boolean isDeleted = iAccountsService.deleteAccount(mobileNumber);
+        if(isDeleted) {
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(new ResponseDto(AccountsConstants.STATUS_200, AccountsConstants.MESSAGE_200));
+        }else{
+            return ResponseEntity
+                    .status(HttpStatus.EXPECTATION_FAILED)
+                    .body(new ResponseDto(AccountsConstants.STATUS_417, AccountsConstants.MESSAGE_417_DELETE));
+        }
+    }
+
+    @Operation(
+            summary = "Get Build information",
+            description = "Get Build information that is deployed into accounts microservice"
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = STATUS_200,
+                    description = MESSAGE_200
+            ),
+            @ApiResponse(
+                    responseCode = STATUS_500,
+                    description = MESSAGE_500,
+                    content = @Content(
+                            schema = @Schema(implementation = ErrorResponseDto.class)
+                    )
+            )
+    }
+    )
+    @GetMapping("/build-info")
+    public ResponseEntity<String> getBuildInfo() {
+        return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(buildVersion);
+    }
+
+    @Operation(
+            summary = "Get Java version",
+            description = "Get Java versions details that is installed into accounts microservice"
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = STATUS_200,
+                    description = MESSAGE_200
+            ),
+            @ApiResponse(
+                    responseCode = STATUS_500,
+                    description = MESSAGE_500,
+                    content = @Content(
+                            schema = @Schema(implementation = ErrorResponseDto.class)
+                    )
+            )
+    }
+    )
+    @GetMapping("/java-version")
+    public ResponseEntity<String> getJavaVersion() {
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(environment.getProperty("JAVA_HOME"));
+    }
+
+    @Operation(
+            summary = "Get Contact Info",
+            description = "Contact Info details that can be reached out in case of any issues"
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    responseCode = STATUS_200,
+                    description = MESSAGE_200
+            ),
+            @ApiResponse(
+                    responseCode = STATUS_500,
+                    description = MESSAGE_500,
+                    content = @Content(
+                            schema = @Schema(implementation = ErrorResponseDto.class)
+                    )
+            )
+    }
+    )
+    @GetMapping("/contact-info")
+    public ResponseEntity<AccountContactInfoDto> getContactInfo() {
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(accountContactInfoDto);
+    }
+
+
+}
