+++ accounts/src/main/java/com/nisbank/accounts/dto/AccountContactInfoDto.java
@@ -1,0 +1,19 @@
+package com.nisbank.accounts.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import java.util.List;
+import java.util.Map;
+
+@ConfigurationProperties(prefix = "accounts")
+@Getter
+@Setter
+public class AccountContactInfoDto {
+
+    private String message;
+    private Map<String, String> contactDetails;
+    private List<String> onCallSupport;
+
+}
