+++ accounts/src/main/java/com/nisbank/accounts/dto/AccountDto.java
@@ -1,0 +1,27 @@
+package com.nisbank.accounts.dto;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Pattern;
+import lombok.Data;
+
+@Data
+@Schema(
+        name = "Accounts",
+        description = "Schema to hold Account information"
+)
+public class AccountDto {
+
+    @NotEmpty(message = "AccountNumber can not be a null or empty")
+    @Pattern(regexp="(^$|[0-9]{10})",message = "AccountNumber must be 10 digits")
+    @Schema(description = "Account Number of Nis Bank account", example = "3454433243")
+    private Long accountNumber;
+
+    @NotEmpty(message = "AccountType can not be a null or empty")
+    @Schema(description = "Account type of Nis Bank account", example = "Savings")
+    private String accountType;
+
+    @NotEmpty(message = "BranchAddress can not be a null or empty")
+    @Schema(description = "Nis Bank branch address", example = "Rabin 12, Herzelya")
+    private String branchAddress;
+}
