+++ accounts/src/main/java/com/nisbank/accounts/entity/Account.java
@@ -1,0 +1,23 @@
+package com.nisbank.accounts.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter @Setter @ToString @AllArgsConstructor @NoArgsConstructor
+public class Account extends  BaseEntity {
+
+    @Column(name="customer_id")
+    private Long customerId;
+
+    @Column(name="account_number")
+    @Id
+    private Long accountNumber;
+
+    @Column(name="account_type")
+    private String accountType;
+
+    @Column(name="branch_address")
+    private String branchAddress;
+
+}
