+++ accounts/src/main/java/com/nisbank/accounts/exception/CustomerAlreadyExistsException.java
@@ -1,0 +1,13 @@
+package com.nisbank.accounts.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.BAD_REQUEST)
+public class CustomerAlreadyExistsException extends RuntimeException {
+
+    public CustomerAlreadyExistsException(String message) {
+        super(message);
+    }
+
+}
