+++ accounts/src/main/java/com/nisbank/accounts/mapper/AccountsMapper.java
@@ -1,0 +1,22 @@
+package com.nisbank.accounts.mapper;
+
+import com.nisbank.accounts.dto.AccountDto;
+import com.nisbank.accounts.entity.Account;
+
+public class AccountsMapper {
+
+    public static AccountDto mapToAccountsDto(Account account, AccountDto accountDto) {
+        accountDto.setAccountNumber(account.getAccountNumber());
+        accountDto.setAccountType(account.getAccountType());
+        accountDto.setBranchAddress(account.getBranchAddress());
+        return accountDto;
+    }
+
+    public static Account mapToAccounts(AccountDto accountDto, Account account) {
+        account.setAccountNumber(accountDto.getAccountNumber());
+        account.setAccountType(accountDto.getAccountType());
+        account.setBranchAddress(accountDto.getBranchAddress());
+        return account;
+    }
+
+}
