+++ accounts/src/main/java/com/nisbank/accounts/mapper/CustomerMapper.java
@@ -1,0 +1,22 @@
+package com.nisbank.accounts.mapper;
+
+import com.nisbank.accounts.dto.CustomerDto;
+import com.nisbank.accounts.entity.Customer;
+
+public class CustomerMapper {
+
+    public static CustomerDto mapToCustomerDto(Customer customer, CustomerDto customerDto) {
+        customerDto.setName(customer.getName());
+        customerDto.setEmail(customer.getEmail());
+        customerDto.setMobileNumber(customer.getMobileNumber());
+        return customerDto;
+    }
+
+    public static Customer mapToCustomer(CustomerDto customerDto, Customer customer) {
+        customer.setName(customerDto.getName());
+        customer.setEmail(customerDto.getEmail());
+        customer.setMobileNumber(customerDto.getMobileNumber());
+        return customer;
+    }
+
+}
