+++ accounts/src/main/java/com/nisbank/accounts/repository/AccountsRepository.java
@@ -1,0 +1,20 @@
+package com.nisbank.accounts.repository;
+
+import com.nisbank.accounts.entity.Account;
+import jakarta.transaction.Transactional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface AccountsRepository extends JpaRepository<Account, Long> {
+
+    Optional<Account> findByCustomerId(Long customerId);
+
+    @Transactional
+    @Modifying
+    void deleteByCustomerId(Long customerId);
+
+}
