+++ accounts/src/main/java/com/nisbank/accounts/service/IAccountsService.java
@@ -1,0 +1,34 @@
+package com.nisbank.accounts.service;
+
+import com.nisbank.accounts.dto.CustomerDto;
+
+public interface IAccountsService {
+
+    /**
+     *
+     * @param customerDto - CustomerDto Object
+     */
+    void createAccount(CustomerDto customerDto);
+
+    /**
+     *
+     * @param mobileNumber - Input Mobile Number
+     * @return Accounts Details based on a given mobileNumber
+     */
+    CustomerDto fetchAccount(String mobileNumber);
+
+    /**
+     *
+     * @param customerDto - CustomerDto Object
+     * @return boolean indicating if the update of Account details is successful or not
+     */
+    boolean updateAccount(CustomerDto customerDto);
+
+    /**
+     *
+     * @param mobileNumber - Input Mobile Number
+     * @return boolean indicating if deletion of Account details is successful or not
+     */
+    boolean deleteAccount(String mobileNumber);
+
+}
