+++ accounts/src/main/java/com/nisbank/accounts/service/impl/AccountsServiceImpl.java
@@ -1,0 +1,115 @@
+package com.nisbank.accounts.service.impl;
+
+import com.nisbank.accounts.constants.AccountsConstants;
+import com.nisbank.accounts.dto.AccountDto;
+import com.nisbank.accounts.dto.CustomerDto;
+import com.nisbank.accounts.entity.Account;
+import com.nisbank.accounts.entity.Customer;
+import com.nisbank.accounts.exception.CustomerAlreadyExistsException;
+import com.nisbank.accounts.exception.ResourceNotFoundException;
+import com.nisbank.accounts.mapper.AccountsMapper;
+import com.nisbank.accounts.mapper.CustomerMapper;
+import com.nisbank.accounts.repository.AccountsRepository;
+import com.nisbank.accounts.repository.CustomersRepository;
+import com.nisbank.accounts.service.IAccountsService;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+import java.util.Random;
+
+@Service
+@AllArgsConstructor
+public class AccountsServiceImpl  implements IAccountsService {
+
+    private AccountsRepository accountsRepository;
+    private CustomersRepository customersRepository;
+
+    /**
+     * @param customerDto - CustomerDto Object
+     */
+    @Override
+    public void createAccount(CustomerDto customerDto) {
+        Customer customer = CustomerMapper.mapToCustomer(customerDto, new Customer());
+        Optional<Customer> optionalCustomer = customersRepository.findByMobileNumber(customerDto.getMobileNumber());
+        if(optionalCustomer.isPresent()) {
+            throw new CustomerAlreadyExistsException("Customer already registered with given mobileNumber"
+                    +customerDto.getMobileNumber());
+        }
+        Customer savedCustomer = customersRepository.save(customer);
+        accountsRepository.save(createNewAccount(savedCustomer));
+    }
+
+    /**
+     * @param customer - Customer Object
+     * @return the new account details
+     */
+    private Account createNewAccount(Customer customer) {
+        Account newAccount = new Account();
+        newAccount.setCustomerId(customer.getCustomerId());
+        long randomAccNumber = 1000000000L + new Random().nextInt(900000000);
+
+        newAccount.setAccountNumber(randomAccNumber);
+        newAccount.setAccountType(AccountsConstants.SAVINGS);
+        newAccount.setBranchAddress(AccountsConstants.ADDRESS);
+        return newAccount;
+    }
+
+    /**
+     * @param mobileNumber - Input Mobile Number
+     * @return Accounts Details based on a given mobileNumber
+     */
+    @Override
+    public CustomerDto fetchAccount(String mobileNumber) {
+        Customer customer = customersRepository.findByMobileNumber(mobileNumber).orElseThrow(
+                () -> new ResourceNotFoundException("Customer", "mobileNumber", mobileNumber)
+        );
+        Account account = accountsRepository.findByCustomerId(customer.getCustomerId()).orElseThrow(
+                () -> new ResourceNotFoundException("Account", "customerId", customer.getCustomerId().toString())
+        );
+        CustomerDto customerDto = CustomerMapper.mapToCustomerDto(customer, new CustomerDto());
+        customerDto.setAccountDto(AccountsMapper.mapToAccountsDto(account, new AccountDto()));
+        return customerDto;
+    }
+
+    /**
+     * @param customerDto - CustomerDto Object
+     * @return boolean indicating if the update of Account details is successful or not
+     */
+    @Override
+    public boolean updateAccount(CustomerDto customerDto) {
+        boolean isUpdated = false;
+        AccountDto accountDto = customerDto.getAccountDto();
+        if(accountDto !=null ){
+            Account account = accountsRepository.findById(accountDto.getAccountNumber()).orElseThrow(
+                    () -> new ResourceNotFoundException("Account", "AccountNumber", accountDto.getAccountNumber().toString())
+            );
+            AccountsMapper.mapToAccounts(accountDto, account);
+            account = accountsRepository.save(account);
+
+            Long customerId = account.getCustomerId();
+            Customer customer = customersRepository.findById(customerId).orElseThrow(
+                    () -> new ResourceNotFoundException("Customer", "CustomerID", customerId.toString())
+            );
+            CustomerMapper.mapToCustomer(customerDto,customer);
+            customersRepository.save(customer);
+            isUpdated = true;
+        }
+        return  isUpdated;
+    }
+
+    /**
+     * @param mobileNumber - Input Mobile Number
+     * @return boolean indicating if deletion of Account details is successful or not
+     */
+    @Override
+    public boolean deleteAccount(String mobileNumber) {
+        Customer customer = customersRepository.findByMobileNumber(mobileNumber).orElseThrow(
+                () -> new ResourceNotFoundException("Customer", "mobileNumber", mobileNumber)
+        );
+        accountsRepository.deleteByCustomerId(customer.getCustomerId());
+        customersRepository.deleteById(customer.getCustomerId());
+        return true;
+    }
+
+}
