+++ accounts/src/main/resources/application.yml
@@ -31,3 +31,6 @@ management:
     web:
       exposure:
         include: "*"
+
+build:
+  version: "v1"
