--- accounts/src/test/java/com/eazybytes/accounts/AccountsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.eazybytes.accounts;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AccountsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
