+++ accounts/src/test/java/com/nisbank/accounts/AccountApplicationTests.java
@@ -1,0 +1,13 @@
+package com.nisbank.accounts;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AccountApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
